refactor: move info components into separate folder
